+++ src/pages/fre/ViewFre.js
@@ -18,6 +18,13 @@ export default function ViewFre() {
 
 
     const deleteRoute = async (id) => {
+        if (
+            window.confirm(
+            `Please confirm deletion`
+            ) === false
+        ) {
+            return;
+        }
         await axios.delete(`http://localhost:3000/fre/${id}`)
         loadFre();
     }
